backend/requirements.txt
Django~=4.2
psycopg2-binary~=2.9
python-dotenv~=1.0
django-environ~=0.10
markdown~=3.4